Fix stream not being closed.
Files.walk creates a new resource stream that is not properly closed without a try with resource statement. I resolved this by merging it into the statement above.

API Note from Java:
This method must be used within a try-with-resources statement or similar control structure to ensure that the stream's open directories are closed promptly after the stream's operations have completed.